requirements.txt
pytest==3.8.0
PyHamcrest==1.9.0
twine
